Allow the auth server to be started if services are Send + Sync
