+++ crates/utils/src/lib.rs
@@ -1,0 +1,19 @@
+use anyhow::{anyhow, Result};
+use tokio::task::JoinHandle;
+
+pub async fn flatten<T>(handle: JoinHandle<Result<T>>) -> Result<T> {
+    match handle.await {
+        Ok(Ok(result)) => Ok(result),
+        Ok(Err(err)) => Err(err),
+        Err(err) => Err(anyhow!("join failed: {err}")),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    #[test]
+    fn it_works() {
+        let result = 2 + 2;
+        assert_eq!(result, 4);
+    }
+}
