+++ services/world/Cargo.toml
@@ -14,6 +14,7 @@ human-panic = "1"
 structopt = "0.3"
 tracing = "0.1"
 azerust-game = { path = "../../crates/game" }
+azerust-utils = { path = "../../crates/utils" }
 azerust-mysql-auth = { path = "../../crates/mysql-auth" }
 azerust-protocol = { path = "../../crates/protocol", features = ["world"] }
 serde = "1"
